backend/requirements/development.txt
# Include base requirements
-r base.txt

# Development tools
django-debug-toolbar>=4.2,<5.0
ipython>=8.18,<9.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-asyncio>=0.21,<1.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code quality
ruff>=0.1,<1.0
mypy>=1.7,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# AI/ML development
torch>=2.1,<3.0
matplotlib>=3.8,<4.0
